OPENVINO BACKEND WORKS!!!
Runs at just over 11 fps :)

When running the model optimiser, make sure to specify the correct scale/mean.
Then, use 1/0 scale/mean for it when passing it to opencv dnn.

At the moment, the OpenVino ignores scale/mean config. Will try to get
it so it uses them, but I don't know how that works when the input is U8?