Detector and counter setup now functional.
Detectors are now initialised when the buscount filter class is
initialised. Need some way of being able to configure which detector to
used, which would need to be placed in buscountd.cpp. Additionally,
other properties should be modifiable. At present defaults are hard
coded, although this should be easy to fix using Glib::Property values.