Changed WorldConfig standards
In/Out lines are defined by a single 'crossing' line.
The crossing line is a vector pointing towards the out side.

All lines and detection boxes are in "normalised" form. i.e. they are scalled between 0 - 1. Means that there is much less need to pass frame sizes around

There can now be an arbituary number of bounding lines

Added a few tests to world config

Still need to pretty up the line drawing code a bit

Seems to be some bug which causes the server to segfault when run alongside the buscounter